appointment-microservice/requirements.txt
pip~=23.3.2
Flask~=3.0.1
Flask-SQLAlchemy~=3.1.1
Flask-Cors~=4.0.0
Flask-JWT-Extended~=4.6.0
Flask-Testing~=0.8.1
SQLAlchemy~=2.0.25
pytest~=7.4.4
python-dotenv~=1.0.0
requests~=2.31.0